Move createGenericQueryMock to web/utils/testing and implement for modifyTag
